+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where clicking on a TinyMCE dialog backdrop wouldn't close the dialog
+
 ## 1.0.0 - 2022-09-15
 
 ### Added